Basic permission issues fixed
