+++ .idea/misc.xml
@@ -11,6 +11,9 @@
         <entry key="../../../../layout/compose-model-1620024952251.xml" value="0.2712962962962963" />
         <entry key="../../../../layout/compose-model-1620104735170.xml" value="0.1" />
         <entry key="../../../../layout/compose-model-1620106810063.xml" value="0.2657407407407407" />
+        <entry key="../../../../layout/compose-model-1620339279446.xml" value="0.335" />
+        <entry key="../../../../layout/compose-model-1620339990411.xml" value="0.1" />
+        <entry key="../../../../layout/compose-model-1620340237375.xml" value="1.0" />
       </map>
     </option>
   </component>
