+++ app/build.gradle.kts
@@ -63,6 +63,7 @@ dependencies {
     implementation("com.google.android.gms:play-services-location:18.0.0")
     implementation ("androidx.compose.runtime:runtime-livedata:1.0.0-beta05")
     implementation ("com.github.atutorme:kotlin-azan:0.0.9")
+    implementation("com.karumi:dexter:6.2.2")
     testImplementation("junit:junit:4.+")
     androidTestImplementation("androidx.test.ext:junit:1.1.2")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.3.0")
